backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.1
django-filter==25.1
drf-yasg==1.21.10
PyMySQL==1.1.0
python-dotenv==1.1.1
django-environ==0.11.2
Pillow==11.3.0
mysqlclient